ai-smart-agriculture/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
Pillow==10.2.0
numpy==1.26.3
scikit-learn==1.4.0
pandas==2.2.0
joblib==1.3.2
tensorflow==2.15.0
aiosqlite==0.19.0
